requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
pulp>=2.8.0
plotly>=5.18.0
rdkit>=2023.9.1
pillow>=10.0.0
groq>=0.9.0
posthog>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pymoo>=0.6.0
numpy>=1.24.0
